+++ src/GRA.Controllers/ViewModel/Join/SinglePageViewModel.cs
@@ -82,5 +82,11 @@ namespace GRA.Controllers.ViewModel.Join
         public SelectList SchoolList { get; set; }
         public SelectList SchoolDistrictList { get; set; }
         public SelectList SchoolTypeList { get; set; }
+
+        public SelectList AskFirstTime { get; set; }
+
+        [DisplayName("Is this your first time participating?")]
+        [Required(ErrorMessage = "Please let us know if this is your first time participating in the program")]
+        public string IsFirstTime { get; set; }
     }
 }