+++ src/GRA.Controllers/ViewModel/Join/Step3ViewModel.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace GRA.Controllers.ViewModel.Join
 {
@@ -30,5 +31,11 @@ namespace GRA.Controllers.ViewModel.Join
         [Phone]
         [MaxLength(15)]
         public string PhoneNumber { get; set; }
+
+        public SelectList AskFirstTime { get; set; }
+
+        [DisplayName("Is this your first time participating?")]
+        [Required(ErrorMessage = "Please let us know if this is your first time participating in the program")]
+        public string IsFirstTime { get; set; }
     }
 }